src/backend/requirements.txt
# AI-miniSOC Backend Requirements

# FastAPI 核心依赖
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.1
pydantic-settings==2.6.0

# 数据库
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP 客户端
httpx==0.27.2
requests==2.32.3

# AI SDK (智谱AI)
zhipuai>=2.1.5

# 环境变量
python-dotenv==1.0.1

# 日志
loguru==0.7.2

# 工具
python-dateutil==2.9.0
